ml/train_classifier.py
"""
Enhanced Attack Classifier with Threat Intelligence

Improvement over 02_attack_classifier.py:
- OLD: Used only timing (time_diff_seconds) → ~70% accuracy
- NEW: Uses timing + threat intelligence → ~85-90%+ accuracy

New Features:
1. abuse_confidence_score - AbuseIPDB reputation (0-100)
2. total_reports - How many times IP was reported
3. is_tor - Boolean flag for Tor exit nodes
4. time_diff_seconds - Original timing feature

Goal: Better distinguish bots from humans using reputation + behavior
"""

import json 
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score, confusion_matrix, classification_report
import matplotlib.pyplot as plt
import seaborn as sns
from pathlib import Path

print("=" * 70)
print("Enhanced Attack Classifier - Threat Intelligence + Timing")
print("=" * 70)

# ===== STEP 1: Load Enriched Data =====
print("\n[1/7] Loading enriched attack data...")
enriched_file = Path('ml/data/enriched_attacks.jsonl')

if not enriched_file.exists():
    print("❌ Error: Run ml/03_threat_enrichment.py first!")
    exit(1)

attacks = []
with open(enriched_file) as f:
    for line in f:
        attacks.append(json.loads(line))

df = pd.DataFrame(attacks)
print(f"✅ Loaded {len(df):,} enriched attacks")

# ===== STEP 2: Extract Features =====
print("\n[2/7] Extracting features from enriched data...")

# Original timing feature (from old model)
df['timestamp'] = pd.to_datetime(df['timestamp'])
df = df.sort_values(['source_ip', 'timestamp'])
df['time_diff'] = df.groupby('source_ip')['timestamp'].diff()
df['time_diff_seconds'] = df['time_diff'].dt.total_seconds()

# NEW: Extract threat intelligence features
def extract_threat_features(row):
    """
    Extract threat intelligence features from enriched attack.
    
    Returns dict with:
    - abuse_score: 0-100 (higher = more malicious)
    - total_reports: Count of abuse reports
    - is_tor: 1 if Tor node, 0 otherwise
    - has_threat_data: 1 if we got threat data, 0 otherwise
    """
    threat_intel = row.get('threat_intelligence', {})
    
    # Check if we successfully got threat data
    if threat_intel.get('status') == 'success':
        threat_data = threat_intel.get('threat_data', {})
        return {
            'abuse_score': threat_data.get('abuseConfidenceScore', 0),
            'total_reports': threat_data.get('totalReports', 0),
            'is_tor': 1 if threat_data.get('isTor', False) else 0,
            'has_threat_data': 1
        }
    else:
        # No threat data (private IP, error, etc.)
        return {
            'abuse_score': 0,
            'total_reports': 0,
            'is_tor': 0,
            'has_threat_data': 0
        }

# Apply feature extraction
threat_features = df.apply(extract_threat_features, axis=1, result_type='expand')
df = pd.concat([df, threat_features], axis=1)

print(f"✅ Extracted threat features:")
print(f"   - Abuse scores: min={df['abuse_score'].min():.0f}, max={df['abuse_score'].max():.0f}, avg={df['abuse_score'].mean():.1f}")
print(f"   - Total reports: min={df['total_reports'].min():.0f}, max={df['total_reports'].max():.0f}, avg={df['total_reports'].mean():.1f}")
print(f"   - Tor nodes: {df['is_tor'].sum():,} IPs")
print(f"   - Has threat data: {df['has_threat_data'].sum():,} / {len(df):,} attacks")

# ===== STEP 3: Create Labels (Bot vs Human) =====
print("\n[3/7] Labeling attacks (bot vs human)...")

def label_attacker(row):
    """
    Label attacker type based on event type.
    
    Logic:
    - Bots: Fast automated scanners (HTTP_REQUEST, CONNECTION)
    - Humans: Interactive sessions (COMMAND_EXECUTION, SSH_LOGIN)
    """
    if row['event_type'] in ['HTTP_REQUEST', 'CONNECTION']:
        return 'bot'
    elif row['event_type'] in ['COMMAND_EXECUTION','SSH_LOGIN','CREDENTIAL_SUBMISSION']:
        return 'human'
    return 'unknown'

df['attacker_type'] = df.apply(label_attacker, axis=1)

# Clean data: remove unknowns and missing timing
df = df[df['attacker_type'] != 'unknown']
df = df.dropna(subset=['time_diff_seconds'])

print(f"✅ Dataset after cleaning: {len(df):,} attacks")
print(f"   - Bot attacks: {len(df[df['attacker_type']== 'bot']):,}")
print(f"   - Human attacks: {len(df[df['attacker_type']== 'human']):,}")

# ===== STEP 4: Prepare Feature Matrix =====
print("\n[4/7] Preparing feature matrices...")

# OLD MODEL: Only timing
X_old = df[['time_diff_seconds']].values
y = df['attacker_type'].values

# NEW MODEL: Timing + Threat Intelligence
X_new = df[['time_diff_seconds', 'abuse_score', 'total_reports', 'is_tor']].values

print(f"✅ Old model features: {X_old.shape[1]} (time_diff_seconds)")
print(f"✅ New model features: {X_new.shape[1]} (time_diff_seconds, abuse_score, total_reports, is_tor)")
print(f"✅ Samples: {len(X_new):,}")

# ===== STEP 5: Train Both Models =====
print("\n[5/7] Training models...")

# Split data
X_old_train, X_old_test, y_train, y_test = train_test_split(X_old, y, test_size=0.2, random_state=42)
X_new_train, X_new_test, _, _ = train_test_split(X_new, y, test_size=0.2, random_state=42)

# Train OLD model (timing only)
print("   Training OLD model (timing only)...")
model_old = LogisticRegression(random_state=42)
model_old.fit(X_old_train, y_train)

# Train NEW model (timing + threat intel)
print("   Training NEW model (timing + threat intel)...")
model_new = RandomForestClassifier(n_estimators=100, random_state=42)
model_new.fit(X_new_train, y_train)

print("✅ Both models trained")

# ===== STEP 6: Evaluate and Compare =====
print("\n[6/7] Evaluating models...")

# OLD model predictions
y_pred_old = model_old.predict(X_old_test)
accuracy_old = accuracy_score(y_test, y_pred_old)

# NEW model predictions
y_pred_new = model_new.predict(X_new_test)
accuracy_new = accuracy_score(y_test, y_pred_new)

print(f"\n{'='*70}")
print(f"📊 MODEL COMPARISON")
print(f"{'='*70}")
print(f"OLD Model (Timing Only):           {accuracy_old*100:.2f}%")
print(f"NEW Model (Timing + Threat Intel): {accuracy_new*100:.2f}%")
print(f"{'='*70}")
print(f"Improvement: +{(accuracy_new - accuracy_old)*100:.2f} percentage points 🚀")
print(f"{'='*70}\n")

# Detailed reports
print("OLD Model Classification Report:")
print(classification_report(y_test, y_pred_old))

print("\nNEW Model Classification Report:")
print(classification_report(y_test, y_pred_new))

# Feature importance (only for RandomForest)
feature_names = ['time_diff_seconds', 'abuse_score', 'total_reports', 'is_tor']
feature_importance = model_new.feature_importances_

print("\n🔍 Feature Importance (What matters most?):")
for name, importance in sorted(zip(feature_names, feature_importance), key=lambda x: x[1], reverse=True):
    print(f"   {name:20s}: {importance:.3f} ({importance*100:.1f}%)")

# ===== STEP 7: Visualizations =====
print("\n[7/7] Creating comparison visualizations...")

charts_dir = Path('ml/charts')
charts_dir.mkdir(exist_ok=True)

# Create figure with subplots
fig, axes = plt.subplots(2, 2, figsize=(14, 10))
fig.suptitle('Enhanced Classifier - Threat Intelligence Impact', fontsize=16, fontweight='bold')

# Plot 1: Accuracy Comparison
ax1 = axes[0, 0]
models = ['OLD\n(Timing Only)', 'NEW\n(Timing + Threat Intel)']
accuracies = [accuracy_old * 100, accuracy_new * 100]
colors = ['#e74c3c', '#2ecc71']
bars = ax1.bar(models, accuracies, color=colors, edgecolor='black', linewidth=2)
ax1.set_ylabel('Accuracy (%)', fontweight='bold')
ax1.set_title('Model Accuracy Comparison')
ax1.set_ylim(0, 100)
ax1.axhline(y=50, color='gray', linestyle='--', alpha=0.5, label='Random guess')
# Add value labels on bars
for bar, acc in zip(bars, accuracies):
    height = bar.get_height()
    ax1.text(bar.get_x() + bar.get_width()/2., height,
            f'{acc:.1f}%', ha='center', va='bottom', fontweight='bold', fontsize=12)
ax1.legend()
ax1.grid(axis='y', alpha=0.3)

# Plot 2: Confusion Matrix - NEW Model
ax2 = axes[0, 1]
cm_new = confusion_matrix(y_test, y_pred_new, labels=['bot', 'human'])
sns.heatmap(cm_new, annot=True, fmt='d', cmap='Greens', 
            xticklabels=['Bot', 'Human'], yticklabels=['Bot', 'Human'], ax=ax2)
ax2.set_xlabel('Predicted', fontweight='bold')
ax2.set_ylabel('Actual', fontweight='bold')
ax2.set_title('NEW Model Confusion Matrix')

# Plot 3: Feature Importance
ax3 = axes[1, 0]
sorted_idx = np.argsort(feature_importance)
ax3.barh(np.array(feature_names)[sorted_idx], feature_importance[sorted_idx], 
         color='#3498db', edgecolor='black')
ax3.set_xlabel('Importance', fontweight='bold')
ax3.set_title('Feature Importance (What Drives Predictions?)')
ax3.grid(axis='x', alpha=0.3)

# Plot 4: Threat Score Distribution by Attacker Type
ax4 = axes[1, 1]
bot_scores = df[df['attacker_type'] == 'bot']['abuse_score']
human_scores = df[df['attacker_type'] == 'human']['abuse_score']
ax4.hist([bot_scores, human_scores], bins=20, label=['Bot', 'Human'], 
         color=['#e74c3c', '#3498db'], alpha=0.7, edgecolor='black')
ax4.set_xlabel('Abuse Confidence Score', fontweight='bold')
ax4.set_ylabel('Count', fontweight='bold')
ax4.set_title('Threat Score Distribution by Attacker Type')
ax4.legend()
ax4.grid(alpha=0.3)

plt.tight_layout()
plt.savefig(charts_dir / 'enhanced_classifier_comparison.png', dpi=150, bbox_inches='tight')
print(f"✅ Saved: enhanced_classifier_comparison.png")
plt.close()

print("\n" + "=" * 70)
print("✅ Enhanced Classifier Complete!")
print("=" * 70)
print(f"\n📈 Key Takeaways:")
print(f"   1. Accuracy improved by {(accuracy_new - accuracy_old)*100:.1f} percentage points")
print(f"   2. Threat intelligence features add significant predictive power")
print(f"   3. Most important feature: {feature_names[np.argmax(feature_importance)]}")
print(f"\n📁 Output:")
print(f"   - Chart: ml/charts/enhanced_classifier_comparison.png")
