requirements.txt
# HoneyPot Dependencies
# Core honeypot
docker>=7.0.0
paramiko>=3.4.0

# Dashboard
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.11.0
eventlet>=0.36.0

# ML & Threat Intelligence
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# PDF Export
xhtml2pdf>=0.2.15

# Utilities
requests>=2.31.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0

